C_MM04有問題

+++ 中文題庫/數學 I/C_MM04.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    int i, n;
+    long long i, n;
     while (cin >> i >> n) {
         cout << i << "+" << n << "=" << i+n << '\n'
              << i << "*" << n << "=" << i*n << '\n'
